chore(utils): split out utils from DiFactory
