+++ lib/AutoLayout.js
@@ -5,6 +5,10 @@ var BpmnModdle = require('bpmn-moddle'),
 var STDDIST = 25;
 const PADDING = 36 / 2
 
+const PADDING_NODE = 'padding_node'
+
+const DiFactoryUtils = require('./DiFactoryUtils')
+
 function AutoLayout() {
   this.moddle = new BpmnModdle();
   this.DiFactory = new DiFactory(this.moddle);
@@ -131,7 +135,7 @@ AutoLayout.prototype._addTreeNode = function (flowElement, flowElements) {
 			while (padParent._getLevel() < maxParentLevel) {
         let oldParent = padParent
         self.tree.removeParentFromNode(node.id, oldParent.id)
-        padParent = self.tree.add(this.nodeCount, `padding node ${this.nodeCount}`, 0, 0)
+        padParent = self.tree.add(this.nodeCount, PADDING_NODE, 0, 0)
         this.nodeCount += 1
         self.tree.addParentToNode(node.id, padParent.id)
         self.tree.addParentToNode(padParent.id, oldParent.id)
@@ -276,40 +280,11 @@ AutoLayout.prototype._layoutElements = function(group, parentDi) {
     height = size.height;
     maxWidth = Math.max(maxWidth, size.width);
 
-
     let treeNode = self.tree.getNodeById(element.nodeId)
     if (treeNode) {
       pos.x = treeNode.XPosition
       pos.y = treeNode.YPosition
     }
-
-    /*
-    if (is(element.$type, 'bpmn:Gateway')) {
-      let incoming = element.get('incoming')
-      let outgoing = element.get('outgoing')
-      if (outgoing.length < incoming.length) {
-        // closing gateway
-        // ... find opening and get y position
-        let source = incoming[0].get('sourceRef')
-        let stack = 0
-        while (source.$type !== element.$type && stack >= 0) {
-          incoming = source.get('incoming')
-          outgoing = source.get('outgoing')
-          source = incoming[0].get('sourceRef')
-          if (source.$type !== element.$type) {
-            if (outgoing.length < incoming.length) {
-              stack += 1
-            } else if (outgoing.length > incoming.length) {
-              stack -= 1
-            }
-          }
-        }
-
-        treeNode = self.tree.getNodeById(source.nodeId)
-        pos.y = treeNode.YPosition
-      }
-    }
-    */
     
     element.bounds = Object.assign({}, size, pos);
     elementDi = createDi('shape', element, pos);
@@ -340,7 +315,10 @@ AutoLayout.prototype._layoutElements = function(group, parentDi) {
 };
 
 AutoLayout.prototype._layoutConnections = function(group, parentDi) {
+  const self = this
   var createDi = this.DiFactory.createBpmnElementDi.bind(this.DiFactory);
+  var createDiWaypoint = this.DiFactory.createDiWaypoint.bind(this.DiFactory);
+  var is = this.DiFactory.is.bind(this.DiFactory);
   var childrenDi = parentDi.get('planeElement');
   var connections = group.connections;
   connections.forEach(function(connection) {
@@ -349,6 +327,47 @@ AutoLayout.prototype._layoutConnections = function(group, parentDi) {
       return childDi.id.includes(connection.id)
     })
     var connectionDi = createDi('connection', connection);
+
+    var sourceRef = connection.get('sourceRef')
+    var targetRef = connection.get('targetRef')
+    var sourceNode = self.tree.getNodeByName(sourceRef.id)
+    var directTarget = false
+    if (sourceNode.children && sourceNode.children.length > 0) {
+      sourceNode.children.forEach(function (node) {
+        if (node.dsc === targetRef.id) {
+          directTarget = true
+        }
+      })
+    }
+
+    if (sourceNode.children && sourceNode.children.length > 0 && !directTarget) {
+      // this connection needs to accommodate padding nodes
+      // check for any padding nodes and go through those coords
+      const oldPoints = connectionDi.waypoint
+      const newPoints = [oldPoints[0]]
+      // get node for sourceRef
+      const pathNodes = getPathToTarget(sourceNode, targetRef.id)
+      pathNodes.pop()
+      pathNodes.shift()
+      let points = []
+      pathNodes.forEach(pathNode => {
+        console.log( `x: ${pathNode.XPosition}, y: ${pathNode.YPosition}`)
+        points.push({ x: pathNode.XPosition, y: pathNode.YPosition})
+      })
+      const lastPoint = oldPoints[oldPoints.length - 1]
+      if (points.length > 0) {
+        points.forEach(point => {
+          newPoints.push(createDiWaypoint(point))
+        })
+        const directions = 'h:v'
+        DiFactoryUtils.getBendpoints(points[points.length - 1], lastPoint, directions).forEach(bendpoint => {
+          newPoints.push(createDiWaypoint(bendpoint))
+        })
+      }
+      newPoints.push(lastPoint)
+      connectionDi.waypoint = newPoints
+    }
+
     if (found) {
       // replace exisiting with latest
       childrenDi.splice(childrenDi.indexOf(found), 1, connectionDi)
@@ -360,6 +379,30 @@ AutoLayout.prototype._layoutConnections = function(group, parentDi) {
 
 
 /////// helpers //////////////////////////////////
+
+function getPathToTarget(sourceNode, targetId) {
+  if (sourceNode === undefined) {
+    return undefined
+  }
+  if (sourceNode.dsc === targetId) {
+    return [sourceNode]
+  }
+  if (sourceNode.children) {
+    let pathToTarget
+    sourceNode.children.forEach(function (source) {
+      if (source.dsc === PADDING_NODE || source.dsc === targetId) {
+        const path = getPathToTarget(source, targetId)
+        if (path) {
+          pathToTarget = path
+        }
+      }
+    })
+    if (pathToTarget) {
+      return [sourceNode, ...pathToTarget]
+    }
+  }
+  return undefined
+}
 
 function getStartEvents(flowElements) {
   return flowElements.filter(function(e) {
