+++ lib/ECOTree.js
@@ -134,7 +134,7 @@ ECOTree = function () {
 	this.config = {
 		iMaxDepth: 100,
 		iLevelSeparation: 50,
-		iSiblingSeparation: 20,
+		iSiblingSeparation: 40,
 		iSubtreeSeparation: 20,
 		topXAdjustment: 0,
 		topYAdjustment: 0
@@ -267,6 +267,7 @@ ECOTree._secondWalk = function (tree, node, level, X, Y, prevSiblings = []) {
 			node.XPosition = node.YPosition;
 			node.YPosition = swapTmp;
 		}
+
 		if (node._getChildrenCount() != 0) {
 			ECOTree._secondWalk(tree, node._getFirstChild(), level + 1, X + node.modifier, Y + maxsizeTmp + tree.config.iLevelSeparation, [...prevSiblings, node.id]);
 		}
